Bug #204784. Adding Tomcat resource agent

+++ rgmanager/src/resources/utils/ra-skelet.sh
@@ -61,6 +61,11 @@ stop_generic()
 
 	read pid < "$pid_file"
 
+	# @todo: PID is not running -> error?
+	if [ ! -d "/proc/$pid_file" ]; then
+		return 0;
+	fi
+                                
 	kill -TERM "$pid"
 
 	if [ $? -ne 0 ]; then
